refactor: split page builder state and middleware
