+++ apps/shop-abc/__tests__/middleware/cache.test.ts
@@ -1,0 +1,11 @@
+jest.mock("@acme/config", () => ({ env: {} }));
+jest.mock("@upstash/redis", () => ({ Redis: jest.fn() }));
+
+import { redis, attempts } from "../../src/middleware/cache";
+
+describe("cache", () => {
+  it("initializes without redis", () => {
+    expect(redis).toBeNull();
+    expect(attempts).toBeInstanceOf(Map);
+  });
+});
