+++ apps/shop-abc/__tests__/middleware/logger.test.ts
@@ -1,0 +1,10 @@
+import { warn } from "../../src/middleware/logger";
+
+describe("logger", () => {
+  it("prefixes scope", () => {
+    const spy = jest.spyOn(console, "warn").mockImplementation(() => {});
+    warn("login", "locked");
+    expect(spy).toHaveBeenCalledWith("[login] locked");
+    spy.mockRestore();
+  });
+});
