+++ apps/shop-abc/src/middleware/logger.ts
@@ -1,0 +1,3 @@
+export function warn(scope: string, message: string) {
+  console.warn(`[${scope}] ${message}`);
+}
